clip-picture-search/requirements.txt
Pillow==10.2.0
transformers==4.37.2
